// src/app/core/model/user-create.ts
export class CreateUser {
  addedBy: any;
  createdAt: any;
  email: string;
  gender: string;
  id: number;
  image: string;
  isActive: boolean;
  isDeleted: boolean;
  licenseId: number;
  mobile: string;
  mobileNo: string;
  name: string;
  role: number;
  token: string;
  updatedAt: any;
  updatedBy: any;
  username: string;
  password?: string;
}